requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.7
billiard==3.3.0.23
celery==3.1.18
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-cors-headers==3.3.0
django-filter==2.1.0
djangorestframework==3.9.2
idna==2.8
kombu==3.0.37
Markdown==3.0.1
pytz==2018.9
redis==2.10.3
requests==2.21.0
sqlparse==0.3.1
urllib3==1.24.1
